+++ benchmarks/material_main.cpp
@@ -107,7 +107,8 @@ bool run_case(std::string_view name, const options& config, Operation&& operatio
         std::chrono::duration<double, std::nano>(clock_type::now() - begin).count();
     samples.push_back(elapsed / static_cast<double>(config.iterations));
   }
-  const auto mean = std::accumulate(samples.begin(), samples.end(), 0.0) / samples.size();
+  const auto mean =
+      std::accumulate(samples.begin(), samples.end(), 0.0) / static_cast<double>(samples.size());
   std::cout << "1," << name << ',' << config.variants << ',' << config.iterations << ','
             << config.samples << ',' << mean << ',' << percentile(samples, 0.50) << ','
             << percentile(samples, 0.95) << ',' << percentile(samples, 0.99) << '\n';
